+++ src/app/(main)/freecell-in-italiano/gioca/page.tsx
@@ -30,6 +30,7 @@ export const metadata: Metadata = {
       es: absoluteUrl('/freecell-en-espanol/jugar'),
       fr: absoluteUrl('/freecell-en-francais/jouer'),
       de: absoluteUrl('/freecell-auf-deutsch/spielen'),
+      pt: absoluteUrl('/freecell-em-portugues/jogar'),
       en: absoluteUrl('/'),
       'x-default': absoluteUrl('/'),
     },