+++ src/app/(main)/page.tsx
@@ -119,6 +119,7 @@ function buildMetadata(): Metadata {
         fr: absoluteUrl('/freecell-en-francais/jouer'),
         de: absoluteUrl('/freecell-auf-deutsch/spielen'),
         it: absoluteUrl('/freecell-in-italiano/gioca'),
+        pt: absoluteUrl('/freecell-em-portugues/jogar'),
         'x-default': absoluteUrl('/'),
       },
     },